Language-Translation-Bot/requirements.txt
streamlit==1.21.0
openai==0.27.6
pillow==9.4.0
streamlit-chat
speechrecognition
pyaudio
gTTS
streamlit-extras